requirements.txt
djangorestframework
django
django-cors-headers
Pillow
markdown
django-filter
